Fix styling Help Dialog
