+++ asreview/webapp/src/Components/index.js
@@ -5,6 +5,7 @@ export { default as DrawerItem } from "./DrawerItem";
 export { default as DrawerItemContainer } from "./DrawerItemContainer";
 export { default as ExportDialog } from "./ExportDialog";
 export { default as Header } from "./Header";
+export { default as HelpDialog } from "./HelpDialog";
 export { default as NavigationDrawer } from "./NavigationDrawer";
 export { default as OpenInNewIconStyled } from "./OpenInNewIconStyled";
 export { default as ProjectImportDialog } from "./ProjectImportDialog";
